requirements.txt
pkg-resources==0.0.0
ply==3.11
pydle==0.9.3
pyhcl==0.3.13
